Updated the build system partially.

+++ build.sh
@@ -1,19 +1,23 @@
 #!/bin/bash
 
-arfile='schematics.com_dev.zip'
-ardest='../schematics.com-deploy'
+arfile='tennisblock_dev.zip'
+ardest='../tennisblock-deploy'
 
 echo "Cleaning up.."
 rm -f $arfile
 mkdir -p $ardest
-rm -rf $ardest/currentVersion
-rm  $ardest/$arfile
+
+if [ -d $ardest ];then
+    rm -rf $ardest/currentVersion
+    rm  $ardest/$arfile
+fi
 
 echo "Building..."
 
 # Build less compiled files and additional javascript libraries.
 build/build.py
 
+exit
 # Collect static files
 python manage.py collectstatic --clear --noinput --ignore eeweb
